+++ src/enums/index.ts
@@ -2,4 +2,3 @@ export * from './amino-acid-charge.js';
 export * from './amino-acid-polarity.js';
 export * from './amino-acid-side-chain-type.js';
 export * from './nucleic-acid-type.js';
-export * from './rna-sub-type.js';